# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gemini Rent Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List all loans/EMIs
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create a loan/EMI record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Loan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Loan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
    delete:
      operationId: deleteLoan
      tags: [loans]
      summary: Delete a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /loans/{id}/payments:
    post:
      operationId: createLoanPayment
      tags: [loans]
      summary: Record an EMI payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanPaymentInput"
      responses:
        "201":
          description: EMI payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanPayment"
  /maintenance:
    get:
      operationId: listMaintenanceRequests
      tags: [maintenance]
      summary: List maintenance requests
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of maintenance requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRequest"
    post:
      operationId: createMaintenanceRequest
      tags: [maintenance]
      summary: Create a maintenance request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceRequestInput"
      responses:
        "201":
          description: Maintenance request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
  /maintenance/{id}:
    patch:
      operationId: updateMaintenanceRequest
      tags: [maintenance]
      summary: Update a maintenance request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
    delete:
      operationId: deleteMaintenanceRequest
      tags: [maintenance]
      summary: Delete a maintenance request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, employee]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, employee]
        createdAt:
          type: string
...
    Expense:
      type: object
      required: [id, title, amount, category, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [repair, utility, tax, insurance, maintenance, salary, other]
        date:
          type: string
        propertyId:
          type: ["integer", "null"]
        propertyName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [title, amount, category, date]
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [repair, utility, tax, insurance, maintenance, salary, other]
        date:
          type: string
        propertyId:
          type: integer
        notes:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [repair, utility, tax, insurance, maintenance, salary, other]
        date:
          type: string
        propertyId:
          type: integer
        notes:
          type: string
    Loan:
      type: object
      required: [id, lenderName, principalAmount, interestRate, emiAmount, startDate, totalMonths, paidMonths, status, createdAt]
      properties:
        id:
          type: integer
        lenderName:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        emiAmount:
          type: number
        startDate:
          type: string
        totalMonths:
          type: integer
        paidMonths:
          type: integer
        remainingAmount:
          type: number
        status:
          type: string
          enum: [active, completed, defaulted]
        propertyId:
          type: ["integer", "null"]
        propertyName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LoanInput:
      type: object
      required: [lenderName, principalAmount, interestRate, emiAmount, startDate, totalMonths]
      properties:
        lenderName:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        emiAmount:
          type: number
        startDate:
          type: string
        totalMonths:
          type: integer
        propertyId:
          type: integer
        notes:
          type: string
    LoanUpdate:
      type: object
      properties:
        lenderName:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        emiAmount:
          type: number
        startDate:
          type: string
        totalMonths:
          type: integer
        status:
          type: string
          enum: [active, completed, defaulted]
        propertyId:
          type: integer
        notes:
          type: string
    LoanPayment:
      type: object
      required: [id, loanId, amount, paymentDate, createdAt]
      properties:
        id:
          type: integer
        loanId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LoanPaymentInput:
      type: object
      required: [amount, paymentDate]
      properties:
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: string
    MaintenanceRequest:
      type: object
      required: [id, title, description, propertyId, priority, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        tenantId:
          type: ["integer", "null"]
        tenantName:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        resolvedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    MaintenanceRequestInput:
      type: object
      required: [title, description, propertyId, priority]
      properties:
        title:
          type: string
        description:
          type: string
        propertyId:
          type: integer
        tenantId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        notes:
          type: string
    MaintenanceRequestUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        notes:
          type: string
...