# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Gemini Rent Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseCategory } from './expenseCategory';

export interface Expense {
  id: number;
  title: string;
  amount: number;
  category: ExpenseCategory;
  date: string;
  /** @nullable */
  propertyId?: number | null;
  /** @nullable */
  propertyName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}