# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Gemini Rent Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseUpdateCategory } from './expenseUpdateCategory';

export interface ExpenseUpdate {
  title?: string;
  amount?: number;
  category?: ExpenseUpdateCategory;
  date?: string;
  propertyId?: number;
  notes?: string;
}