# lib/api-zod/src/generated/types/maintenanceRequestUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Gemini Rent Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceRequestUpdatePriority } from './maintenanceRequestUpdatePriority';
import type { MaintenanceRequestUpdateStatus } from './maintenanceRequestUpdateStatus';

export interface MaintenanceRequestUpdate {
  title?: string;
  description?: string;
  priority?: MaintenanceRequestUpdatePriority;
  status?: MaintenanceRequestUpdateStatus;
  notes?: string;
}